+++ templates/login.html
@@ -70,7 +70,7 @@
                   <h4>Log in</h4>
                 </div>
 
-                <form id="login-form" action="{{ url_for('api_login') }}" method="post" data-sitekey="{{reCAPTCHA_site_key}}">
+                <form id="login-form" action="{{ url_for('auth.api_login') }}" method="post" data-sitekey="{{reCAPTCHA_site_key}}">
                   {{ login_form.hidden_tag() }}
                   <div class='row'>
                     <div class='input-field col s12'>
@@ -114,7 +114,7 @@
                 <div class="row">
                   <h4>Create your account</h4>
                 </div>
-                <form id="register-form" action="{{ url_for('register') }}" method="post" data-sitekey="{{reCAPTCHA_site_key}}">
+                <form id="register-form" action="{{ url_for('auth.register') }}" method="post" data-sitekey="{{reCAPTCHA_site_key}}">
                   {{ register_form.hidden_tag() }}
                   <div class="row">
                     <div class='input-field col s6'>
@@ -171,7 +171,7 @@
         <p class="flow-text">It's okay! It happens to the best of us.</p>
         <p class="flow-text">Enter the email associated with your account below and we'll send you a link to reset your password.</p>
         <div class="divider"></div>
-        <form id="forgot-password-form" action="{{ url_for('forgot_password') }}" method="post" data-sitekey="{{reCAPTCHA_site_key}}">
+        <form id="forgot-password-form" action="{{ url_for('auth.forgot_password') }}" method="post" data-sitekey="{{reCAPTCHA_site_key}}">
           {{ login_form.hidden_tag() }}
           <div class="row">
             <div class="row">
