+++ test/unit/test_routes.py
@@ -1,0 +1,32 @@
+import pytest
+from flask import Flask
+
+# Import your Flask application
+from budgeteer import create_app
+
+@pytest.fixture
+def app():
+    """Create and configure a new instance of the Flask application for testing."""
+    app = create_app()  # Replace `create_app` with your function that returns a Flask app instance
+    app.testing = True  # Set app to testing mode
+    return app
+
+@pytest.fixture
+def client(app):
+    """Create a test client for the Flask application."""
+    return app.test_client()
+
+def test_index(client):
+    """Test the index route."""
+    response = client.get('/')  # Make a GET request to the index route
+    assert response.status_code == 200  # Check that the response status code is 200 OK
+    assert b'Hello, World!' in response.data  # Check that the response contains the expected content
+
+def test_post_request(client):
+    """Test a POST request route."""
+    # Make a POST request to the desired route
+    response = client.post('/your_post_route', data={'key': 'value'})
+    assert response.status_code == 200  # Check response status code
+    assert response.json == {'success': True}  # Check JSON response
+
+# Add more tests as needed
